fix context + dotenv
