+++ frontend/src/.env
@@ -1,1 +1,1 @@
-PUBLIC_API_DOMAIN=http://localhost:3000/api
+GATEWAY_API_URL=http://localhost:3000/api
